backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
httpx==0.26.0
PyJWT==2.8.0
pymongo==4.6.1
motor==3.3.2
torch==2.1.2
sentence-transformers==2.2.2
scikit-learn==1.4.0
pandas==2.1.4
numpy==1.26.3
passlib[bcrypt]==1.7.4
